+++ androidApp/src/main/java/com/rld/justlisten/android/ui/playlistscreen/components/Header.kt
@@ -1,0 +1,18 @@
+package com.rld.justlisten.android.ui.playlistscreen.components
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.rld.justlisten.android.ui.theme.typography
+
+@Composable
+fun Header(text: String, modifier: Modifier = Modifier) {
+    Text(
+        text = text,
+        style = typography.h5.copy(fontWeight = FontWeight.ExtraBold),
+        modifier = modifier.padding(start = 8.dp, end = 4.dp, bottom = 8.dp, top = 24.dp)
+    )
+}
